@@ -8,7 +8,7 @@ class Task::Notifications::CompletionReminderEmail < Mutations::Command
       next unless user.email.present?
       token = Token.task_completion.create! context: { user_id: user.id, task_id: task.id }
 
-      TaskMailer.delay.task_completion_reminder(task.id, user.id, token.code)
+      TaskMailer.task_completion_reminder(task.id, user.id, token.code).deliver_later
     end
   end
 end
