@@ -9,7 +9,7 @@ class Task::Notifications::ReminderEmail < Mutations::Command
       Token.task_confirmation.for_user_id(user.id).update_all(active: false)
       token    = Token.task_confirmation.create! context: { user_id: user.id, task_id: task.id }
 
-      TaskMailer.delay.task_reminder(task.id, user.id, token.code)
+      TaskMailer.task_reminder(task.id, user.id, token.code).deliver_later
     end
   end
 end
