@@ -1,6 +1,6 @@
 class SecureHashSerializer
   def initialize key=Rails.application.secrets.secret_key_base
-    @crypt = ActiveSupport::MessageEncryptor.new(key)
+    @crypt = ActiveSupport::MessageEncryptor.new(truncate_key(key))
   end
 
   def load encrypted
@@ -17,5 +17,15 @@ class SecureHashSerializer
       string = YAML.dump(hash)
       @crypt.encrypt_and_sign(string)
     end
+  end
+
+  private
+
+  # Running Ruby >= 2.4, we get an "ArgumentError: key must be 32 bytes" when trying to use the full, 128 bytes
+  # long secret key base for OpenSSL::Cipher, so we need to truncate it. This is exactly what Ruby < 2.4 did before,
+  # they silently discarded everything after the first 32 bytes of the key. Details here:
+  # https://github.com/rails/rails/pull/25192
+  def truncate_key(key)
+    key[0, 32]
   end
 end
